add: hud: input + audio device
- Microphone volume in HUD
- Audio Input/Output device update in HUD
- Initial code for Power Status

+++ MewNotch/Utils/Helpers/Brightness/Brightness.h
@@ -12,7 +12,7 @@
 
 @interface Brightness : NSObject
 + (Brightness*) sharedInstance;
-@property (getter=brightness, setter=setBrightness:) double brightness;
+@property (getter=brightness, setter=setBrightness:) float brightness;
 @end
 
 extern NSString *BrightnessNotification;
